Add "Create User" option
When this option is enabled, the plugin retains the previous behavior of
allowing Joomla! to create a user account for users who successfully
authenticate with the mailbox server and do not already have a Joomla!
account.  When this option is disabled, users without an existing
Joomla! account are rejected.

This will allow administrators greater control of which users are
allowed to login to Joomla! (rather than allowing all email users
access).

Signed-off-by: Kevin Locke <klocke@digitalenginesoftware.com>