Translate plugin name
Update the plugin name to match the other authentication plugins shipped
with Joomla!.  Provide a translation for the name in a new .sys.ini
translation file.  This file only includes plugin name and description
translations.  It is loaded by the extension manager instead of the
normal .ini file to reduce translation load times.

Signed-off-by: Kevin Locke <klocke@digitalenginesoftware.com>