Prepare for release of 1.0.6
Update the versions and changelog in preparation for releasing the
current version.

Version 1.0.5 caused a significant amount of confusion, both due to a
typo in the tag which labeled it 1.0.4 and because there was not a
release for Joomla! 1.5.  1.0.6 is being released without any changes as
a way of moving past the mess that was 1.0.5.

Signed-off-by: Kevin Locke <klocke@digitalenginesoftware.com>