Support domain in username
This commit adds support for including a domain in the login username
and optionally in the resulting Joomla! username.  This fixes the
usability issue noted in #12 that users may enter their full email
address as their login username.

It additionally adds support for (optionally) using the whole email
address as the Joomla! username.  This can be necessary for use cases
where the mail server hosts multiple domains with users allowed to
access Joomla!.

Fixes: #12

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>