Clarify license is GNU GPL 2 or later
The previous license statement was ambiguous about whether the plugin
was "GPLv2 or later" or "GPLv2 only".  Since any Joomla! code for which
this plugin may be a derivative work was declared to be "GPLv2 or later"
in joomla/joomla-cms@de4a567e there is a clear path forward.

I hereby license all of my copyrightable works in this project under the
terms of the GNU General Public License as published by the Free
Software Foundation; either version 2 of the License, or (at your
option) any later version.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>