composer.json: Create composer.json with project info
Although I don't have any current plans to publish this project using
Composer, define the project information in composer.json in case it is
useful to anyone.

I'm not sure if the joomla-extension type has any significance, but it
is used by a few existing packages[1] and seems likely to have the
correct meaning.

1.  https://packagist.org/?type=joomla-extension

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>