Improve translation key prefixes
To disambiguate the translation keys in this plugin from other plugins
and Joomla! core, and to match the convention of other plugins, use the
plugin name as a prefix.  Shorten `_AUTHENTICATION_` to `_AUTH_` to make
the keys more usable (and match the existing convention in .sys.ini).

Also remove the `PARAM` prefix from keys which have it, since nothing
about the string is parameter-specific and may be used in other
contexts.

Based on changes by Manuel Pérez Ayala.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>