mailbox.xml: Wrap long <field> elements
<field> elements on lines which are significantly longer than 100
characters are difficult to read and changes in diffs are obscured.
Wrap each attribute onto its own line for legibility and diff-ability.
Put closing > on its own line for extra content space on the line and to
minimize diff size when changing last attribute.  (This matches the
style of several core plugins like GMail).

No code changes in this commit.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>