mailbox.php: Improve imap_open call formatting
Improve the formatting of the call to imap_open to make it easier to
read.  Also add a comment about the $n_retries parameter.  In my
testing, NNTP failed to authenticate when $n_retries was explicitly 0
(and tried several times when it was unspecified).  1 seems to work well
for all protocols (and doesn't retry any in my testing).

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>