Add support for multiple mail servers
This commit is based on a [multi-server version] of this plugin created
by Manuel Pérez Ayala.  I have already integrated several of the fixes
and the es-ES language files into the master branch.  This commit adds
the core changes necessary for multi-server support, but omits the
repeatable tab layout for now.

Fixes: #3

[multi-server version]: https://github.com/kevinoid/auth-mailbox-joomla/issues/3#issuecomment-449068699

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>